+++ src/components/Cart.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const Cart = () => {
+  return <div>Cart</div>;
+};
+
+export default Cart;
