+++ src/components/Footer.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const Footer = () => {
+  return <div>Footer</div>;
+};
+
+export default Footer;
